Move TransactionType to api module
